v.0.2.5
Changed the specification when URL is described in the code section. The add-on itself reads the script from the URL and executes the loaded code. Previously, the URL was converted to code that injects a script tag that specifies the url in the src property and executed.  
If you want to access the variables/functions/objects, etc. created by the page script, specify wrapCodeInScriptTag in the //options directive.